# интерфейсы/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Кейс 02: регистрация, оплата и доступ",
    "description": "Исполняемая демонстрация проектируемого контракта. Не принимает настоящие платежи. Тренировочные данные API не принимает.",
    "version": "1.0.0"
  },
  "paths": {
    "/v1/auth/challenges": {
      "post": {
        "tags": [
          "Вход"
        ],
        "summary": "Запросить код входа или регистрации",
        "operationId": "requestChallenge",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChallengeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Запрос принят; результат следует уточнить",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeOut"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный запрос",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Необходим действующий сеанс",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Запись не найдена или недоступна",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт состояния или повторного запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Недопустимые входные данные",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Превышена частота запросов",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Сервис временно недоступен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/me": {
      "get": {
        "tags": [
          "Вход"
        ],
        "summary": "Получить свою учётную запись",
        "operationId": "getAccount",
        "responses": {
          "200": {
            "description": "Операция выполнена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountOut"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный запрос",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Необходим действующий сеанс",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Запись не найдена или недоступна",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт состояния или повторного запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Недопустимые входные данные",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Превышена частота запросов",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Сервис временно недоступен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountOut": {
        "properties": {
          "accountId": {
            "type": "string",
            "format": "uuid",
            "title": "Идентификатор учётной записи",
            "description": "Идентификатор учётной записи. Идентификатор поля в формате обмена: accountId."
          },
          "email": {
            "type": "string",
            "title": "Адрес электронной почты",
            "description": "Адрес электронной почты. Идентификатор поля в формате обмена: email."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "accountId",
          "email"
        ],
        "title": "Учётная запись"
      },
...
      "ChallengeIn": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "minLength": 3,
            "pattern": "^[^\\s@]+@[^\\s@]+\\.[^\\s@]+$",
            "title": "Адрес электронной почты",
            "description": "Адрес электронной почты. Идентификатор поля в формате обмена: email."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email"
        ],
        "title": "Запрос кода входа"
      },
      "ChallengeOut": {
        "properties": {
          "challengeId": {
            "type": "string",
            "format": "uuid",
            "title": "Идентификатор запроса кода входа",
            "description": "Идентификатор запроса кода входа. Идентификатор поля в формате обмена: challengeId."
          },
          "expiresInSeconds": {
            "type": "integer",
            "title": "Срок действия в секундах",
            "description": "Срок действия в секундах. Идентификатор поля в формате обмена: expiresInSeconds."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "challengeId",
          "expiresInSeconds"
        ],
        "title": "Выданный запрос кода"
      },
...
      "Error": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Код",
            "description": "Код. Идентификатор поля в формате обмена: code."
          },
          "message": {
            "type": "string",
            "title": "Сообщение для пользователя",
            "description": "Сообщение для пользователя. Идентификатор поля в формате обмена: message."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Ошибка операции"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Ключ сеанса передаётся в заголовке Authorization: Bearer <ключ>. Значение не публикуется в журнале."
      }
    }
  }
}